// packages/frontend/src/generated/types/StorageResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * @description Response DTO for the `storage` entity.
 */
export interface StorageResponse {
  /**
   * @type string,null
   */
  capacity?: string | null
  /**
   * @type string
   */
  commonName: string
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type boolean
   */
  isTypeSpecific: boolean
  /**
   * @type string,null
   */
  longName?: string | null
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string,null, uuid
   */
  productTypeId?: string | null
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
  /**
   * @type string, uuid
   */
  warehouseId: string
}